+++ packages/components/src/components/block-section/block-section.e2e.ts
@@ -1,6 +1,6 @@
 import { newE2EPage, E2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it } from "vitest";
-import { themed } from "../../tests/commonTests";
+
 import { html } from "../../../support/formatting";
 import { CSS } from "./resources";
 
@@ -186,51 +186,4 @@ it("should emit expanded/collapsed events when toggled", async () => {
   expect(await item.getProperty("expanded")).toBe(false);
   expect(expandSpy).toHaveReceivedEventTimes(1);
   expect(collapseSpy).toHaveReceivedEventTimes(1);
-});
-
-describe("theme", () => {
-  describe("default", () => {
-    themed(
-      html`
-        <calcite-block-section text="Planes" expanded icon-end="pen" icon-start="pen" text="a block-section">
-          <p>Block section content</p>
-        </calcite-block-section>
-      `,
-      {
-        "--calcite-block-section-border-color": {
-          targetProp: "borderBlockEndColor",
-        },
-        "--calcite-block-section-background-color": {
-          shadowSelector: `.${CSS.toggle}`,
-          targetProp: "backgroundColor",
-        },
-        "--calcite-block-section-header-text-color": [
-          {
-            targetProp: "color",
-          },
-        ],
-        "--calcite-block-section-text-color": [
-          { shadowSelector: `.${CSS.chevronIcon}`, targetProp: "color" },
-          { shadowSelector: `.${CSS.iconStart}`, targetProp: "color" },
-          { shadowSelector: `.${CSS.iconEnd}`, targetProp: "color" },
-        ],
-        "--calcite-block-section-text-color-hover": [
-          {
-            shadowSelector: `.${CSS.toggle}`,
-            targetProp: "color",
-            state: "hover",
-          },
-          {
-            shadowSelector: `.${CSS.chevronIcon}`,
-            targetProp: "color",
-            state: "hover",
-          },
-        ],
-        "--calcite-block-section-content-space": {
-          shadowSelector: `.${CSS.content}`,
-          targetProp: "paddingBlock",
-        },
-      },
-    );
-  });
 });