+++ packages/components/src/components/inline-editable/inline-editable.e2e.ts
@@ -1,6 +1,6 @@
 import { E2EPage, newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { beforeEach, describe, expect, it } from "vitest";
-import { labelable, themed } from "../../tests/commonTests";
+import { labelable } from "../../tests/commonTests";
 import { html } from "../../../support/formatting";
 import type { Input } from "../input/input";
 import { findAll, getElementRect, toElementHandle } from "../../tests/utils/puppeteer";
@@ -328,57 +328,4 @@ describe("has controls", () => {
       );
     });
   });
-});
-
-describe("theme", () => {
-  themed("calcite-inline-editable", {
-    "--calcite-inline-editable-background-color-hover": {
-      shadowSelector: `.${CSS.wrapper}`,
-      state: "hover",
-      targetProp: "backgroundColor",
-    },
-    "--calcite-inline-editable-background-color": {
-      shadowSelector: `.${CSS.wrapper}`,
-      targetProp: "backgroundColor",
-    },
-  });
-  themed(
-    html`<calcite-inline-editable controls editing-enabled>
-      <calcite-input />
-    </calcite-inline-editable>`,
-    {
-      "--calcite-inline-editable-button-corner-radius": [
-        {
-          shadowSelector: `.${CSS.enableEditingButton}`,
-          targetProp: "--calcite-button-corner-radius",
-        },
-        {
-          shadowSelector: `.${CSS.cancelEditingButton}`,
-          targetProp: "--calcite-button-corner-radius",
-        },
-        {
-          shadowSelector: `.${CSS.confirmChangesButton}`,
-          targetProp: "--calcite-button-corner-radius",
-        },
-      ],
-      "--calcite-inline-editable-button-loader-color": {
-        shadowSelector: `.${CSS.confirmChangesButton}`,
-        targetProp: "--calcite-button-loader-color",
-      },
-      "--calcite-inline-editable-button-text-color": [
-        {
-          shadowSelector: `.${CSS.enableEditingButton}`,
-          targetProp: "--calcite-button-text-color",
-        },
-        {
-          shadowSelector: `.${CSS.cancelEditingButton}`,
-          targetProp: "--calcite-button-text-color",
-        },
-        {
-          shadowSelector: `.${CSS.confirmChangesButton}`,
-          targetProp: "--calcite-button-text-color",
-        },
-      ],
-    },
-  );
 });