+++ packages/components/src/components/input-message/input-message.e2e.ts
@@ -1,7 +1,6 @@
 import { newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it, beforeEach } from "vitest";
-import { themed } from "../../tests/commonTests";
-import { html } from "../../../support/formatting";
+
 import { StatusIconDefaults } from "./interfaces";
 import { CSS } from "./resources";
 
@@ -122,47 +121,6 @@ describe("when icon prop is provided", () => {
         requestedIcon = await iconEl.getAttribute("icon");
         expect(requestedIcon).toEqual("view-hide");
       });
-    });
-  });
-});
-
-describe("theme", () => {
-  describe("status", () => {
-    describe("invalid", () => {
-      themed(html`<calcite-input-message icon status="invalid">Message</calcite-input-message>`, {
-        "--calcite-input-message-icon-color": {
-          shadowSelector: `.${CSS.inputMessageIcon}`,
-          targetProp: "color",
-        },
-      });
-    });
-    describe("valid", () => {
-      themed(html`<calcite-input-message icon status="valid">Message</calcite-input-message>`, {
-        "--calcite-input-message-icon-color": {
-          shadowSelector: `.${CSS.inputMessageIcon}`,
-          targetProp: "color",
-        },
-      });
-    });
-    describe("idle", () => {
-      themed(html`<calcite-input-message icon status="idle">Message</calcite-input-message>`, {
-        "--calcite-input-message-icon-color": {
-          shadowSelector: `.${CSS.inputMessageIcon}`,
-          targetProp: "color",
-        },
-      });
-    });
-  });
-
-  describe("deprecated", () => {
-    themed(html`<calcite-input-message icon status="invalid">Message</calcite-input-message>`, {
-      "--calcite-input-message-spacing-value": {
-        targetProp: "marginBlockStart",
-      },
-      "--calcite-ui-icon-color": {
-        shadowSelector: `.${CSS.inputMessageIcon}`,
-        targetProp: "color",
-      },
     });
   });
 });