+++ packages/components/src/components/link/link.e2e.ts
@@ -1,7 +1,6 @@
 import { newE2EPage, E2EElement } from "@arcgis/lumina-compiler/puppeteerTesting";
-import { describe, expect, it } from "vitest";
-import { themed } from "../../tests/commonTests";
-import { html } from "../../../support/formatting";
+import { expect, it } from "vitest";
+
 import { CSS } from "./resources";
 
 it("sets download attribute on internal anchor", async () => {
@@ -149,15 +148,4 @@ it('renders with an icon-start and icon-end and role="button"', async () => {
   expect(iconStart).not.toBeNull();
   expect(elementAsLink).toEqualAttribute("role", "button");
   expect(iconEnd).not.toBeNull();
-});
-
-describe("theme", () => {
-  describe("default", () => {
-    themed(html` <calcite-link href="#" icon-start="banana" icon-end="information">link</calcite-link> `, {
-      "--calcite-link-text-color": {
-        shadowSelector: "a",
-        targetProp: "color",
-      },
-    });
-  });
 });