+++ packages/components/src/components/pagination/pagination.browser.e2e.tsx
@@ -8,6 +8,7 @@ import {
   hidden,
   renders,
   t9n,
+  themed,
 } from "../../tests/commonTests/browser";
 import { CSS } from "./resources";
 
@@ -55,4 +56,93 @@ describe("renders", () => {
 
 describe("translation support", () => {
   t9n(() => mount("calcite-pagination"));
+});
+
+describe("theme", () => {
+  describe("default", () => {
+    themed(() => mount(<calcite-pagination page-size="100" start-item="1" total-items="1200" />), {
+      "--calcite-pagination-color": [
+        {
+          shadowSelector: `.${CSS.chevron}`,
+          targetProp: "color",
+        },
+        {
+          shadowSelector: `.${CSS.page}:not(.${CSS.selected})`,
+          targetProp: "color",
+        },
+        {
+          shadowSelector: `.${CSS.ellipsis}`,
+          targetProp: "color",
+        },
+      ],
+    });
+  });
+  describe("hover", () => {
+    themed(() => mount(<calcite-pagination page-size="100" start-item="1" total-items="1200" />), {
+      "--calcite-pagination-color-hover": [
+        {
+          shadowSelector: `.${CSS.chevron}:not(.${CSS.disabled})`,
+          targetProp: "color",
+          state: "hover",
+        },
+        {
+          shadowSelector: `.${CSS.page}`,
+          targetProp: "color",
+          state: "hover",
+        },
+      ],
+      "--calcite-pagination-color-border-hover": {
+        shadowSelector: `.${CSS.page}:not(.${CSS.selected})`,
+        targetProp: "borderBlockEndColor",
+        state: "hover",
+      },
+      "--calcite-pagination-icon-color-background-hover": {
+        shadowSelector: `.${CSS.chevron}:not(.${CSS.disabled})`,
+        targetProp: "backgroundColor",
+        state: "hover",
+      },
+    });
+  });
+  describe("active", () => {
+    themed(() => mount(<calcite-pagination page-size="100" start-item="1" total-items="1200" />), {
+      "--calcite-pagination-color-hover": [
+        {
+          shadowSelector: `.${CSS.chevron}:not(.${CSS.disabled})`,
+          targetProp: "color",
+          state: { press: { attribute: "class", value: `${CSS.chevron}` } },
+        },
+        {
+          shadowSelector: `.${CSS.page}`,
+          targetProp: "color",
+          state: { press: { attribute: "class", value: `${CSS.page}` } },
+        },
+      ],
+      "--calcite-pagination-background-color": [
+        {
+          shadowSelector: `.${CSS.page}:not(.${CSS.selected})`,
+          targetProp: "backgroundColor",
+          state: { press: { attribute: "class", value: `${CSS.page}` } },
+        },
+        {
+          shadowSelector: `.${CSS.chevron}:not(.${CSS.disabled})`,
+          targetProp: "backgroundColor",
+          state: { press: { attribute: "class", value: `${CSS.chevron}` } },
+        },
+      ],
+    });
+  });
+  describe("selected", () => {
+    themed(() => mount(<calcite-pagination page-size="100" start-item="1" total-items="1200" />), {
+      "--calcite-pagination-color-hover": {
+        shadowSelector: `.${CSS.page}`,
+        targetProp: "color",
+        state: "focus",
+      },
+      "--calcite-pagination-color-border-active": {
+        shadowSelector: `.${CSS.page}`,
+        targetProp: "borderBlockEndColor",
+        state: "focus",
+      },
+    });
+  });
 });